refactor: move ContentstackLivePreview to new folder
